【feature】：add ResourceManager
