+++ Engine/Source/Runtime/Function/Meta/Object.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include "Function/EngineInit.h"
 #include "Rtti.h"
 #include "Rttimacro.h"
 #include "ObjName.h"
